[merb-gen] Component generator support is now working.
The generators now work for both application and component generators.
The non-ORM model and controller have been ported to the new system,
with some improvements.  The controller may now be nested arbitrarily deeply.
The rspec and test_unit merb_controller_test generators are now not
appropriate for use with resource_controller.  There should be
another generator made for this.