Added bundling for require_js and require_css; track cached bundles in-memory
The helper methods require_js and require_css now accept options to be merged
with the include_required_js and include_required_css options. So :bundle and
in the case of css :media options work as expected.

A list of cached bundles is kept in memory instead of looking for an existing
file (which can be out of date). The in-memory tracking of bundles implies
that bundles are regenerated whenever the app (re)starts. In other words, at
deployment too. (What about starting up multiple instances?)