Leave a notice on libraries under lib/ autoloading in init.rb (see details).
Bootloader in merb-core sets :lib type path globbing to nil,
meaning no globbing for /lib directory. So libraries
organized in a way

 - lib/
    magicwand/
      lib/
			  magicwand.rb

will not play nice with autoload. Using Merb.push_path
for lib in init.rb with **/*.rb globbing does the trick but
changing this in core causes spec suite to blow up.
Until this investigated
and resolved, we leave a comment in generated init.rb.