[merb-core] Fixed generated POST body in MultipartRequestHelper
Previously, Merb::Test::MultipartRequestHelper::Post
generated an incorrect body for forms which have more
than one non-file parameter. In addition to that, the
array wasn't correctly encoded.

Thx to Sandi Metz and Micheal Opitz for this patch!

I only altered it to not put the specs into spec10
but spec instead.