[merb-core] Fixed multipart specs (including spec10)
All merb-core specs pass again.

params[:tempfile] is of kind Tempfile and not File
(as was specced previously).

The only reason this broken spec passed earlier,
was due to a bug in rspec <= 1.2.3 that allowed
Tempfile.new('random').should be_a_kind_of(File)
In rspec > 1.2.4 running the same will throw an error.